Remove old project, rename new project
