+++ TwitchEbooks.sln
@@ -3,11 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.30225.117
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TwitchEbooks", "TwitchEbooks\TwitchEbooks.csproj", "{09CBCDE8-6671-4D4F-8483-C77081FF9B98}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TwitchEbooks.Database", "TwitchEbooks.Database\TwitchEbooks.Database.csproj", "{7DF55FC1-90BE-4DBC-87CC-D440134333BF}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TwitchEbooks2", "TwitchEbooks2\TwitchEbooks2.csproj", "{B4BEAAC0-7EA0-4DA5-A360-7E53E2F306AA}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TwitchEbooks", "TwitchEbooks\TwitchEbooks.csproj", "{00548FA2-756C-4B1E-96F6-F6A8D54AB92E}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,18 +13,14 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{09CBCDE8-6671-4D4F-8483-C77081FF9B98}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{09CBCDE8-6671-4D4F-8483-C77081FF9B98}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{09CBCDE8-6671-4D4F-8483-C77081FF9B98}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{09CBCDE8-6671-4D4F-8483-C77081FF9B98}.Release|Any CPU.Build.0 = Release|Any CPU
 		{7DF55FC1-90BE-4DBC-87CC-D440134333BF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{7DF55FC1-90BE-4DBC-87CC-D440134333BF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7DF55FC1-90BE-4DBC-87CC-D440134333BF}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{7DF55FC1-90BE-4DBC-87CC-D440134333BF}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B4BEAAC0-7EA0-4DA5-A360-7E53E2F306AA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B4BEAAC0-7EA0-4DA5-A360-7E53E2F306AA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B4BEAAC0-7EA0-4DA5-A360-7E53E2F306AA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B4BEAAC0-7EA0-4DA5-A360-7E53E2F306AA}.Release|Any CPU.Build.0 = Release|Any CPU
+		{00548FA2-756C-4B1E-96F6-F6A8D54AB92E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{00548FA2-756C-4B1E-96F6-F6A8D54AB92E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{00548FA2-756C-4B1E-96F6-F6A8D54AB92E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{00548FA2-756C-4B1E-96F6-F6A8D54AB92E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
