+++ TwitchEbooks/Handlers/GenerateMessageNotificationHandler.cs
@@ -2,10 +2,10 @@
 using Microsoft.Extensions.Logging;
 using System.Threading;
 using System.Threading.Tasks;
-using TwitchEbooks2.Infrastructure;
-using TwitchEbooks2.Models.Notifications;
+using TwitchEbooks.Infrastructure;
+using TwitchEbooks.Models.Notifications;
 
-namespace TwitchEbooks2.Handlers
+namespace TwitchEbooks.Handlers
 {
     public class GenerateMessageNotificationHandler : INotificationHandler<GenerateMessageNotification>
     {
