+++ TwitchEbooks/Handlers/IgnoreUserNotificationHandler.cs
@@ -7,10 +7,10 @@ using System.Threading;
 using System.Threading.Tasks;
 using TwitchEbooks.Database;
 using TwitchEbooks.Database.Models;
-using TwitchEbooks2.Infrastructure;
-using TwitchEbooks2.Models.Notifications;
+using TwitchEbooks.Infrastructure;
+using TwitchEbooks.Models.Notifications;
 
-namespace TwitchEbooks2.Handlers
+namespace TwitchEbooks.Handlers
 {
     public class IgnoreUserNotificationHandler : INotificationHandler<IgnoreUserNotification>
     {
