+++ TwitchEbooks/Handlers/JoinNotificationHandler.cs
@@ -1,29 +1,30 @@
 ﻿using MediatR;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using TwitchEbooks.Database;
 using TwitchEbooks.Database.Models;
-using TwitchEbooks2.Infrastructure;
-using TwitchEbooks2.Models.Notifications;
+using TwitchEbooks.Infrastructure;
+using TwitchEbooks.Models.Notifications;
 using TwitchLib.Api;
 using TwitchLib.Client;
 
-namespace TwitchEbooks2.Handlers
+namespace TwitchEbooks.Handlers
 {
-    public class LeaveNotificationHandler : INotificationHandler<LeaveNotification>
+    public class JoinNotificationHandler : INotificationHandler<JoinNotification>
     {
-        private readonly ILogger<LeaveNotificationHandler> _logger;
+        private readonly ILogger<JoinNotificationHandler> _logger;
         private readonly IDbContextFactory<TwitchEbooksContext> _contextFactory;
         private readonly IMarkovChainService _chainService;
         private readonly TwitchAPI _twitchApi;
         private readonly TwitchClient _twitchClient;
 
-        public LeaveNotificationHandler(
-            ILogger<LeaveNotificationHandler> logger,
+        public JoinNotificationHandler(
+            ILogger<JoinNotificationHandler> logger,
             IDbContextFactory<TwitchEbooksContext> contextFactory,
             IMarkovChainService chainService,
             TwitchAPI twitchApi,
@@ -36,31 +37,29 @@ namespace TwitchEbooks2.Handlers
             _twitchClient = twitchClient;
         }
 
-        public async Task Handle(LeaveNotification notification, CancellationToken cancellationToken)
+        public async Task Handle(JoinNotification notification, CancellationToken cancellationToken)
         {
             var context = _contextFactory.CreateDbContext();
-            var channel = context.Channels.FirstOrDefault(c => c.Id == notification.ChannelId);
-            if (channel is null)
+            if (context.Channels.Any(c => c.Id == notification.ChannelId))
                 return;
 
-            // disconnect from channel on Twitch
             // make sure a user exists on Twitch first
             var usersResponse = await _twitchApi.Helix.Users.GetUsersAsync(ids: new List<string> { notification.ChannelId.ToString() });
-            if (usersResponse.Users.Length == 1)
-            {
-                var user = usersResponse.Users[0];
-                _twitchClient.LeaveChannel(user.Login);
-            }
+            if (usersResponse.Users.Length != 1)
+                throw new Exception("No Twitch user could be found by that ID");
+            var user = usersResponse.Users[0];
 
-            // remove channel from database
+            // add channel to database
             context.Channels.Add(new TwitchChannel
             {
                 Id = notification.ChannelId
             });
             await context.SaveChangesAsync(cancellationToken);
 
-            // also remove their Markov chain
-            _chainService.RemoveChainForChannel(notification.ChannelId);
+            // connect to channel on Twitch client
+            _twitchClient.JoinChannel(user.Login);
+            // also create a Markov chain
+            await _chainService.AddOrUpdateChainForChannel(notification.ChannelId);
         }
     }
 }
