--- TwitchEbooks/Models/Events/PurgeWordRequestReceivedEventArgs.cs
@@ -1,8 +1,0 @@
-﻿namespace TwitchEbooks.Models.Events
-{
-    public class PurgeWordRequestReceivedEventArgs
-    {
-        public uint ChannelId { get; set; }
-        public string Word { get; set; }
-    }
-}
