--- TwitchEbooks/Models/Events/TokensRefreshedEventArgs.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using TwitchEbooks.Database.Models;
-
-namespace TwitchEbooks.Models.Events
-{
-    public class TokensRefreshedEventArgs : EventArgs
-    {
-        public TokensRefreshedEventArgs(UserAccessToken tokens) : base()
-        {
-            NewTokens = tokens;
-        }
-
-        public UserAccessToken NewTokens { get; set; }
-    }
-}
