+++ TwitchEbooks/Models/TwitchSettings.cs
@@ -5,5 +5,6 @@
         public string ClientId { get; set; }
         public string ClientSecret { get; set; }
         public string RedirectUri { get; set; }
+        public string BotUsername { get; set; }
     }
 }
