--- TwitchEbooks/Properties/launchSettings.json
@@ -1,13 +1,0 @@
-{
-  "profiles": {
-    "TwitchEbooks": {
-      "commandName": "Project",
-      "environmentVariables": {
-        "DOTNET_ENVIRONMENT": "Development"
-      }
-    },
-    "Docker": {
-      "commandName": "Docker"
-    }
-  }
-}
