+++ TwitchEbooks/appsettings.json
@@ -12,6 +12,7 @@
   "Twitch": {
     "ClientId": "${Twitch.ClientId}",
     "ClientSecret": "${Twitch.ClientSecret}",
-    "RedirectUri": "${Twitch.RedirectUri}"
+    "RedirectUri": "${Twitch.RedirectUri}",
+    "BotUsername": "${Twitch.BotUsername}"
   }
 }
