--- TwitchEbooks2/Models/TwitchSettings.cs
@@ -1,10 +1,0 @@
-﻿namespace TwitchEbooks2.Models
-{
-    public class TwitchSettings
-    {
-        public string ClientId { get; set; }
-        public string ClientSecret { get; set; }
-        public string RedirectUri { get; set; }
-        public string BotUsername { get; set; }
-    }
-}
