--- TwitchEbooks2/appsettings.json
@@ -1,18 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft": "Warning",
-      "Microsoft.Hosting.Lifetime": "Information"
-    }
-  },
-  "ConnectionStrings": {
-    "PostgreSQL": "${ConnectionStrings.PostgreSQL}"
-  },
-  "Twitch": {
-    "ClientId": "${Twitch.ClientId}",
-    "ClientSecret": "${Twitch.ClientSecret}",
-    "RedirectUri": "${Twitch.RedirectUri}",
-    "BotUsername": "${Twitch.BotUsername}"
-  }
-}
